requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sqlalchemy==1.4.52
pymysql==1.1.1
pydantic==1.10.19
python-dotenv==1.0.1
openai==1.30.5
httpx==0.27.2
pytest==8.3.5
